+++ src/roiextractors/extractors/caiman/caimansegmentationextractor.py
@@ -5,6 +5,7 @@ Classes
 CaimanSegmentationExtractor
     A class for extracting segmentation from CaImAn output.
 """
+
 from pathlib import Path
 
 try: