+++ src/roiextractors/extractors/hdf5imagingextractor/__init__.py
@@ -11,4 +11,5 @@ Classes
 Hdf5ImagingExtractor
     An imaging extractor for HDF5.
 """
+
 from .hdf5imagingextractor import Hdf5ImagingExtractor