+++ src/roiextractors/extractors/tiffimagingextractors/tiffimagingextractor.py
@@ -5,6 +5,7 @@ Classes
 TiffImagingExtractor
     A TIFF imaging extractor for TIFF files.
 """
+
 from pathlib import Path
 from typing import Optional
 from warnings import warn