+++ src/roiextractors/testing.py
@@ -1,4 +1,5 @@
 """Testing utilities for the roiextractors package."""
+
 from collections.abc import Iterable
 from typing import Tuple, Optional
 from warnings import warn
@@ -425,6 +426,11 @@ def assert_get_frames_return_shape(imaging_extractor: ImagingExtractor):
     frame_idxs = np.array([0, 1])
     frames_with_array = imaging_extractor.get_frames(frame_idxs=frame_idxs, channel=0)
     assert_msg = "get_frames does not work correctly with frame_idxs=np.arrray([0, 1])"
+    assert frames_with_array.shape == (2, image_size[0], image_size[1]), assert_msg
+
+    frame_idxs = [0, 2]
+    frames_with_array = imaging_extractor.get_frames(frame_idxs=frame_idxs, channel=0)
+    assert_msg = "get_frames does not work correctly with frame_idxs=[0, 2]"
     assert frames_with_array.shape == (2, image_size[0], image_size[1]), assert_msg